feat(component): menu
Move menu from inside Wrapper
