+++ package.json
@@ -59,6 +59,7 @@
     "react-redux": "^7.2.4",
     "react-router-dom": "^5.2.0",
     "react-toastify": "^7.0.4",
+    "react-transition-group": "^4.4.2",
     "react-use": "^17.2.4",
     "remark-gfm": "^3.0.1",
     "styled-components": "^5.3.0",
@@ -98,6 +99,7 @@
     "@types/react-phone-number-input": "^3.0.11",
     "@types/react-router-dom": "^5.1.7",
     "@types/react-select-country-list": "^2.2.0",
+    "@types/react-transition-group": "^4.4.4",
     "@types/redux-mock-store": "^1.0.3",
     "@types/styled-components": "^5.1.10",
     "@types/webfontloader": "^1.6.33",
