+++ src/components/scrolltotop/index.ts
@@ -1,0 +1,3 @@
+import ScrollToTop from './ScrollToTop';
+
+export default ScrollToTop;
