+++ src/index.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
 import GraphqlProvider from './providers/graphql/GraphqlProvider';
 import GraphqlClient from './api/graphql/GraphqlClient';
-import App from './App';
+import App from './app';
 import reportWebVitals from './reportWebVitals';
 import './styles/css/animate.min.css';
 import './styles/css/noscript.css';
